refactor: essay and report 통합
