+++ src/queries/essayQueryOptions.ts
@@ -2,6 +2,8 @@ import { DetailParams, ListParams } from "../types/params";
 
 import getEssayDetail from "../api/essays/getEssayDetail";
 import getEssayList from "../api/essays/getEssayList";
+import getReportDetail from "../api/essays/getReportDetail";
+import getReportList from "../api/essays/getReportList";
 import { queryOptions } from "@tanstack/react-query";
 
 const essayQueryOptions = {
@@ -20,6 +22,16 @@ const essayQueryOptions = {
     queryOptions({
       queryKey: ["dashboard", "essays"],
       queryFn: () => getEssayDetail(params),
+    }),
+  getReportList: (params: ListParams) =>
+    queryOptions({
+      queryKey: ["report", "list", params],
+      queryFn: () => getReportList(params),
+    }),
+  getReportDetail: (params: DetailParams) =>
+    queryOptions({
+      queryKey: ["report", params.id],
+      queryFn: () => getReportDetail(params),
     }),
 };
